+++ protected/controllers/PersonalController.php
@@ -112,12 +112,22 @@ class PersonalController extends Controller
     public function actionDelete($id)
     {
         if (Yii::app()->request->isPostRequest) {
-// we only allow deletion via POST request
-            $this->loadModel($id)->delete();
-
-// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-            if (!isset($_GET['ajax']))
+            try{
+                $grupo_activo = Analisis::model()->findByAttributes(['personal_id'=>$id]);
+                if(!is_null($grupo_activo)){
+                    throw new Exception("Error. Este personal ya posee las asociaciones realizadas");
+                }
+                $this->loadModel($id)->delete();
+                $data = "Se elimino correctamente el personal";
+                echo CJSON::encode($data);
+            }catch (Exception $exception){
+                $data = $exception->getMessage();
+                echo CJSON::encode($data);
+                die();
+            }
+            if (!isset($_GET['ajax'])) {
                 $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+            }
         } else
             throw new CHttpException(400, 'Invalid request. Please do not repeat this request again.');
     }
